perf[Mock]: set responseFake to mock-server.js
https://github.com/PanJiaChen/vue-element-admin/pull/2966
